refactor: Add .amp to final few AMP specific files
