+++ dotcom-rendering/src/types/article.amp.tsx
