+++ dotcom-rendering/src/web/AMPAttributes.amp.ts
