+++ dotcom-rendering/src/web/components/TopMetaLiveblog.amp.tsx
@@ -5,7 +5,7 @@ import { getAgeWarning } from '../../lib/age-warning';
 import { getSoleContributor } from '../../lib/byline';
 import { neutralBorder, pillarPalette_DO_NOT_USE } from '../../lib/pillars';
 import { getSharingUrls } from '../../lib/sharing-urls';
-import type { AMPArticleModel } from '../../types/ampArticleModel';
+import type { AMPArticleModel } from '../../types/article.amp';
 import { Byline } from './Byline.amp';
 import { MainMedia } from './MainMedia.amp';
 import { TopMetaExtras } from './TopMetaExtras.amp';
